0.1.1 - 删除测试数据 - 完善编辑词典功能 - 优化部分代码

+++ app/controllers/edit_toplevel/word_dic_toplevel.py
@@ -20,8 +20,25 @@ def input_data(word_dic_toplevel):
         number += 1
     return True
 
+# 导出至词典数据前置函数：检查词名列表是否有空值
+def refresh_list(word_dic_toplevel):
+    start_iid = word_dic_toplevel.show_word_dic_toplevel.insert('', 0)  # 先在开头创建一个元素，存到变量中
+    temp_iid = word_dic_toplevel.show_word_dic_toplevel.next(start_iid)  # 循环用到的变量，将开头元素赋值进去
+    word_dic_toplevel.show_word_dic_toplevel.delete(start_iid)  # 随后把开头元素删除
+    while temp_iid:
+        word_name = word_dic_toplevel.show_word_dic_toplevel.set(temp_iid, column="word_name")  # 获取词名
+        if not word_name:
+            messagebox.showerror("错误", "当前词典中词名存在空值，请修改")
+            return False
+        next_temp_iid = word_dic_toplevel.show_word_dic_toplevel.next(temp_iid)  # 利用当前元素iid找下一个元素的iid
+        temp_iid = next_temp_iid  # 开启下一个循环
+    return True
+
 # 将编辑词典窗口数据导出至词典数据
 def output_data(word_dic_toplevel):
+    if not refresh_list(word_dic_toplevel):
+        return False
+
     word_dic_toplevel.main_window.word_dic_datas.delete_all_word_dic() # 首先删除data中全部元素
     start_iid = word_dic_toplevel.show_word_dic_toplevel.insert('', 0)  # 先在开头创建一个元素，存到变量中
     temp_iid = word_dic_toplevel.show_word_dic_toplevel.next(start_iid)  # 循环用到的变量，将开头元素赋值进去
@@ -175,25 +192,68 @@ def refresh_entry(word_dic_toplevel):
         word_dic_toplevel.word_frequency_entry.delete(first=0, last="end")
         word_dic_toplevel.word_class_entry.delete(first=0, last="end")
 
+# 输入框变化事件
 def entry_change(word_dic_toplevel):
-    word_dic_toplevel.select_result_iid = word_dic_toplevel.show_word_dic_toplevel.selection()
+    word_dic_toplevel.select_result_iid = word_dic_toplevel.show_word_dic_toplevel.selection() # 获取当前选中项并赋值
 
-    word_name = word_dic_toplevel.word_name_var.get()
+    word_name = word_dic_toplevel.word_name_var.get() # 获取文本框内容
     word_frequency = word_dic_toplevel.word_frequency_var.get()
     word_class = word_dic_toplevel.word_class_var.get()
 
-    if word_dic_toplevel.last_word_name and word_dic_toplevel.last_word_frequency and word_dic_toplevel.last_word_class and word_name and word_frequency and word_class and word_dic_toplevel.show_word_dic_toplevel.selection() == word_dic_toplevel.last_select_result_iid:
+    if word_dic_toplevel.last_word_name and word_dic_toplevel.last_word_frequency and word_dic_toplevel.last_word_class and word_name and word_frequency and word_class and word_dic_toplevel.show_word_dic_toplevel.selection() == word_dic_toplevel.last_select_result_iid: # 如果选项没有变化但是内容发生了变化
         if word_name != word_dic_toplevel.last_word_name or word_frequency != word_dic_toplevel.last_word_frequency or word_class != word_dic_toplevel.last_word_class:
-            not_saved(word_dic_toplevel)
+            not_saved(word_dic_toplevel) # 将保存状态置为未保存状态
 
-    if len(word_dic_toplevel.select_result_iid) == 1:
-        word_dic_toplevel.show_word_dic_toplevel.set(word_dic_toplevel.select_result_iid[0], column='word_name', value=word_name)
+    if len(word_dic_toplevel.select_result_iid) == 1: # 如果只选中一项
+        word_dic_toplevel.show_word_dic_toplevel.set(word_dic_toplevel.select_result_iid[0], column='word_name', value=word_name) # 改变表中内容
         word_dic_toplevel.show_word_dic_toplevel.set(word_dic_toplevel.select_result_iid[0], column='word_frequency', value=word_frequency)
         word_dic_toplevel.show_word_dic_toplevel.set(word_dic_toplevel.select_result_iid[0], column='word_class', value=word_class)
 
-        word_dic_toplevel.last_select_result_iid = word_dic_toplevel.select_result_iid
+        word_dic_toplevel.last_select_result_iid = word_dic_toplevel.select_result_iid # 将上一个选项内容赋值
         word_dic_toplevel.last_word_name = word_name
         word_dic_toplevel.last_word_frequency = word_frequency
         word_dic_toplevel.last_word_class = word_class
     else:
         return
+
+def create_word_dic(word_dic_toplevel):
+    # 获取当前选中的索引，如果没有或者多选的情况下就在尾部添加
+    if len(word_dic_toplevel.select_result_iid) == 1:
+        select_result_index = word_dic_toplevel.show_word_dic_toplevel.index(word_dic_toplevel.show_word_dic_toplevel.next(word_dic_toplevel.select_result_iid[0])) # 获取选中元素下一个的索引
+        new_word_dic = word_dic_toplevel.show_word_dic_toplevel.insert('', select_result_index, values=(0, "", "", ""))
+    else:
+        new_word_dic = word_dic_toplevel.show_word_dic_toplevel.insert('', "end", values=(0, "", "", ""))
+    sort_number(word_dic_toplevel) # 对序号进行排序
+    word_dic_toplevel.show_word_dic_toplevel.selection_set(new_word_dic)  # 象征性地选中新建项
+    word_dic_toplevel.select_result_iid = (new_word_dic, ) # 选中新建项
+    word_dic_toplevel.show_word_dic_toplevel.see(new_word_dic)  # 视图转到新项
+    not_saved(word_dic_toplevel) # 未保存状态
+    refresh_entry(word_dic_toplevel) # 刷新输入框
+
+# 查找分词结果
+def search_result(word_dic_toplevel):
+    search_entry = word_dic_toplevel.search_entry.get() # 获取输入框中的文本
+    if not search_entry:
+        messagebox.showinfo("提示", "当前搜索框内容为空！")
+        return False
+
+    start_iid = word_dic_toplevel.show_word_dic_toplevel.insert('', 0) # 先在开头创建一个元素，存到变量中
+    temp_iid = word_dic_toplevel.show_word_dic_toplevel.next(start_iid) # 循环用到的变量，将开头元素赋值进去
+    word_dic_toplevel.show_word_dic_toplevel.delete(start_iid)  # 随后把开头元素删除
+
+    search_result_select = [] # 列表，用来记录查找结果的iid
+    while temp_iid:
+        temp_result = word_dic_toplevel.show_word_dic_toplevel.item(temp_iid) # 返回当前元素item
+        if search_entry in temp_result['values'][1]: # 如果当前词名中含有搜索文本
+            search_result_select.append(temp_iid) # 将对应iid加入列表
+        next_temp_iid = word_dic_toplevel.show_word_dic_toplevel.next(temp_iid) # 利用当前元素iid找下一个元素的iid
+        temp_iid = next_temp_iid # 开启下一个循环
+
+    if not search_result_select: # 如果没找到词典
+        messagebox.showerror("错误", "未找到词典结果")
+        return False
+    else:
+        word_dic_toplevel.select_result_iid = tuple(search_result_select) # 将查找结果iid给全局变量，选中元素元组
+        word_dic_toplevel.show_word_dic_toplevel.selection_set(word_dic_toplevel.select_result_iid) # 象征性选中查找结果
+        word_dic_toplevel.show_word_dic_toplevel.see(word_dic_toplevel.select_result_iid[0]) # 视图转到查找结果的第一项
+        return True
